+++ packages/gen-typescript-declarations/src/test/fixtures/custom/module-a.js
@@ -10,9 +10,3 @@ export const notASupportedAnalyzerFeature = 0;
 export {NotResolvable} from './module-b.js';
 
 import {PrivateClass} from './module-b.js';
-
-/**
- * @param {!AutoImportClass1} p
- * @return {!AutoImportClass2}
- */
-export function autoImportFunction(p) { }