+++ packages/gen-typescript-declarations/CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   referenced as keys in the `autoImport` config option.
 * Do not `import` or `export` features with `private` visibility.
 * Display warnings generated during the analysis phase.
+* `autoImport` now supports bare-module specifiers. Local files must now
+  begin with `.`.
 
 ## [1.4.0] - 2018-07-25
 - Support for ES module imports and exports.